requirements.txt
Flask
flask-migrate
Flask-SQLAlchemy
Jinja2
Markdown
Pelican
pydantic
python-dateutil
python-dotenv
python-slugify
pyyaml
gunicorn
